Use react query in webhooks-status, refetch it when configu updated
